Add some comments and separate `UnitInputFormatter` protocol and its decimal implementation to separate files.
