+++ WooCommerce/Classes/ViewRelated/ReusableViews/UnitInputTableViewCell.swift
@@ -1,35 +1,5 @@
 import UIKit
 
-struct DecimalInputFormatter: UnitInputFormatter {
-    private let numberFormatter: NumberFormatter = {
-        let numberFormatter = NumberFormatter()
-        numberFormatter.numberStyle = .decimal
-        return numberFormatter
-    }()
-
-    func isValid(input: String) -> Bool {
-        guard input.isEmpty == false else {
-            // Allows empty input to be replaced by 0.
-            return true
-        }
-        return numberFormatter.number(from: input) != nil
-    }
-
-    func format(input text: String?) -> String {
-        guard let text = text, text.isEmpty == false else {
-            return "0"
-        }
-
-        let formattedText = text.replacingOccurrences(of: "^0+([1-9]+)", with: "$1", options: .regularExpression)
-        return formattedText
-    }
-}
-
-protocol UnitInputFormatter {
-    func isValid(input: String) -> Bool
-    func format(input: String?) -> String
-}
-
 struct UnitInputViewModel {
     let title: String
     let unit: String
@@ -38,6 +8,8 @@ struct UnitInputViewModel {
     let onInputChange: ((_ input: String?) -> Void)?
 }
 
+/// Displays a title, an editable text field for user input and the unit of the text field value.
+///
 final class UnitInputTableViewCell: UITableViewCell {
     @IBOutlet private weak var titleLabel: UILabel!
     @IBOutlet private weak var inputAndUnitStackView: UIStackView!
